requirements.txt
amqp==5.2.0
billiard==4.2.1
blinker==1.8.2
cachelib==0.9.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.1
dnspython==2.7.0
filelock==3.13.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Mail==0.10.0
fsspec==2024.2.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
gunicorn==23.0.0
httplib2==0.22.0
huggingface-hub==0.25.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kombu==5.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.3
packaging==24.1
pika==1.3.2
pillow==10.2.0
pluggy==1.5.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pymongo==4.10.1
pyparsing==3.1.4
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.1.1
regex==2024.9.11
requests==2.32.3
rsa==4.9
safetensors==0.4.5
six==1.16.0
sympy==1.12
tokenizers==0.20.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.9.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
